# services/control-plane/docs/openapi.yaml
openapi: 3.0.3
info:
  title: RunIt API
  description: |
    API for managing projects, runs, secrets, and deployments in RunIt.

    ## Authentication
    All endpoints (except health checks and metrics) require a valid Bearer token in the Authorization header.
    Use an API key or JWT token.

    ## Rate Limiting
    - Authenticated users: 120 requests per minute
    - Anonymous users: 60 requests per minute
  version: 0.1.0
  contact:
    name: RunIt
    url: https://runit.dev
  license:
    name: MIT
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.runit.dev
    description: Production
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth JWT token
  schemas:
...
    Context:
      type: object
      required:
        - id
        - name
        - url
        - created_at
      properties:
        id:
          type: string
          format: uuid
          description: Context ID
        name:
          type: string
          description: Context name (alphanumeric, hyphens, underscores)
          example: "my-api-docs"
        url:
          type: string
          format: uri
          description: Source URL the context was fetched from
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        fetched_at:
          type: string
          format: date-time
        size:
          type: integer
          description: Size in bytes
    FetchContextRequest:
      type: object
      required:
        - url
        - name
      properties:
        url:
          type: string
          format: uri
          description: URL to fetch context data from
        name:
          type: string
          pattern: "^[a-zA-Z0-9_-]+$"
          description: Context name (letters, numbers, hyphens, underscores)

    # ==================== Deploy ====================
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: "Project not found"
        code:
          type: string
          description: Error code for programmatic handling
          example: "PROJECT_NOT_FOUND"
        details:
          type: string
          description: Additional error details
    ValidationError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Validation error message
          example: "Missing required fields: name, source_type"
        received:
          type: string
          description: Value that was received (for debugging)
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
          examples:
            missingFields:
              value:
                error: "Missing required fields: name, source_type"
            invalidFormat:
              value:
                error: "Invalid secret key format. Use UPPERCASE_SNAKE_CASE (A-Z, 0-9, _)"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Authentication required"
    Forbidden:
      description: Not authorized to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Not authorized"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            projectNotFound:
              value:
                error: "Project not found"
            runNotFound:
              value:
                error: "Run not found"
            versionNotFound:
              value:
                error: "Version not found"
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Rate limit exceeded. Try again in 60 seconds."
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per minute
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in the current window
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"
paths:
...
  /v1/projects/{id}/context:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Project ID

    post:
      tags: [Context]
      summary: Add context from URL
      description: Fetches context data from a URL and stores it for the project
      operationId: createContext
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FetchContextRequest'
      responses:
        '201':
          description: Context created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  data:
                    type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      tags: [Context]
      summary: List project contexts
      description: Returns all context entries for a project (metadata only, no data)
      operationId: listContexts
      responses:
        '200':
          description: List of contexts
          content:
            application/json:
              schema:
                type: object
                required:
                  - contexts
                properties:
                  contexts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Context'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/projects/{id}/context/{cid}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Project ID
      - name: cid
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Context ID

    get:
      tags: [Context]
      summary: Get context data
      description: Returns the full context data for a specific context entry
      operationId: getContext
      responses:
        '200':
          description: Context data
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  project_id:
                    type: string
                    format: uuid
                  name:
                    type: string
                  url:
                    type: string
                  data:
                    type: object
                  created_at:
                    type: string
                    format: date-time
                  updated_at:
                    type: string
                    format: date-time
                  fetched_at:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags: [Context]
      summary: Refresh context from URL
      description: Re-fetches context data from the original URL
      operationId: refreshContext
      responses:
        '200':
          description: Context refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  data:
                    type: object
                  updated_at:
                    type: string
                    format: date-time
                  fetched_at:
                    type: string
                    format: date-time
        '400':
          description: Cannot refresh context without URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [Context]
      summary: Delete context
      description: Permanently deletes a context entry
      operationId: deleteContext
      responses:
        '200':
          description: Context deleted
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                properties:
                  success:
                    type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ==================== Share ====================
...